Prevent nil values in Client:Set
