+++ cli/api/mock_store.go
@@ -1,0 +1,96 @@
+package api
+
+import (
+	"github.com/vsco/dcdr/cli/api/stores"
+	"github.com/vsco/dcdr/cli/models"
+)
+
+type MockStore struct {
+	Item  *stores.KVByte
+	Items stores.KVBytes
+	Err   error
+}
+
+func NewMockStore(ft *models.Feature, err error) (ms *MockStore) {
+	bts, _ := ft.ToJson()
+
+	ms = &MockStore{
+		Err: err,
+	}
+
+	if ft != nil {
+		kvb := stores.KVBytes{
+			&stores.KVByte{
+				Key:   ft.Key,
+				Bytes: bts,
+			},
+		}
+
+		ms.Item = kvb[0]
+		ms.Items = kvb
+	}
+
+	return
+}
+
+func (ms *MockStore) List(prefix string) (stores.KVBytes, error) {
+	return ms.Items, ms.Err
+}
+
+func (ms *MockStore) Get(key string) (*stores.KVByte, error) {
+	return ms.Item, ms.Err
+}
+
+func (ms *MockStore) Set(key string, bts []byte) error {
+	return ms.Err
+}
+
+func (ms *MockStore) Delete(key string) error {
+	return ms.Err
+}
+
+func (ms *MockStore) Put(key string, bts []byte) error {
+	return ms.Err
+}
+
+type MockRepo struct {
+	error   error
+	sha     string
+	exists  bool
+	enabled bool
+}
+
+func (mr *MockRepo) Clone() error {
+	return mr.error
+}
+
+func (mr *MockRepo) Commit(bts []byte, msg string) error {
+	return mr.error
+}
+
+func (mr *MockRepo) Create() error {
+	return mr.error
+}
+
+func (mr *MockRepo) Exists() bool {
+	return mr.exists
+}
+
+func (mr *MockRepo) Enabled() bool {
+	return mr.enabled
+}
+
+func (mr *MockRepo) Push() error {
+	return mr.error
+}
+
+func (mr *MockRepo) Pull() error {
+	return mr.error
+}
+
+func (mr *MockRepo) CurrentSha() (string, error) {
+	return mr.sha, mr.error
+}
+
+func (mr *MockRepo) Init() {
+}
